package into a library
